# Salus/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Salus API",
        "description": "Documentation de l'API Salus.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/api/symptoms": {
            "get": {
                "tags": [
                    "Symptoms"
                ],
                "summary": "Lister les symptomes",
                "operationId": "0f91203163deecb9f56d775d5cecacad",
                "responses": {
                    "200": {
                        "description": "Liste des symptomes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des symptomes"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Symptom"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifie"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Symptoms"
                ],
                "summary": "Ajouter un symptome",
                "operationId": "184bd58268c64ace214b8a8ccf16e240",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "severity",
                                    "date_recorded"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Fievre"
                                    },
                                    "severity": {
                                        "type": "string",
                                        "example": "moderate"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Fievre depuis 2 jours"
                                    },
                                    "date_recorded": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-03-26"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Prise de paracetamol"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Symptome ajoute",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Symptome ajoute"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Symptom"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifie"
                    },
                    "422": {
                        "description": "Validation echouee"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/symptoms/{id}": {
            "get": {
                "tags": [
                    "Symptoms"
                ],
                "summary": "Voir un symptome",
                "operationId": "3f27648e3e6552102392fd1a48c6d141",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail du symptome",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detail du symptome"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Symptom"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifie"
                    },
                    "404": {
                        "description": "Symptome introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Symptoms"
                ],
                "summary": "Mettre a jour un symptome",
                "operationId": "718cbe6bb658d69aaa712a291b76b764",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Fievre"
                                    },
                                    "severity": {
                                        "type": "string",
                                        "example": "low"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Fievre en baisse"
                                    },
                                    "date_recorded": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-03-26"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Hydratation"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Symptome mis a jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Symptome mis a jour"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Symptom"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifie"
                    },
                    "404": {
                        "description": "Symptome introuvable"
                    },
                    "422": {
                        "description": "Validation echouee"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Symptoms"
                ],
                "summary": "Supprimer un symptome",
                "operationId": "a2bf9854002083810d297bee14850b6e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Symptome supprime",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Symptome supprime"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifie"
                    },
                    "404": {
                        "description": "Symptome introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Symptom": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Fievre"
                    },
                    "severity": {
                        "type": "string",
                        "example": "moderate"
                    },
                    "description": {
                        "type": "string",
                        "example": "Fievre depuis 2 jours"
                    },
                    "date_recorded": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-03-26"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Prise de paracetamol"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Entrez le token au format: Bearer <token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}